@@ -1,55 +1,26 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
-    <PreserveCompilationContext>true</PreserveCompilationContext>
-    <AssemblyName>Web</AssemblyName>
-    <OutputType>Exe</OutputType>
-    <PackageId>WixToolset.Web</PackageId>
-    <Description>WiX Toolset Website</Description>
-    <RuntimeFrameworkVersion>1.1.1</RuntimeFrameworkVersion>
-    <PackageTargetFallback>$(PackageTargetFallback);net46</PackageTargetFallback>
+    <TargetFramework>net5.0</TargetFramework>
+
+    <TinySiteRootFolder>..\Site</TinySiteRootFolder>
+    <TinySitePublishFolder>$(MSBuildProjectDirectory)\wwwroot</TinySitePublishFolder>
+    <SitemapRootUrl>https://www.wixtoolset.com/</SitemapRootUrl>
   </PropertyGroup>
 
   <ItemGroup>
-    <Content Update="Static/**" CopyToPublishDirectory="Never" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.8" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Cookies" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Rewrite" Version="1.0.4" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.AzureAppServicesIntegration" Version="1.0.4" />
-    <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="1.1.6" />
-    <PackageReference Include="Microsoft.AspNetCore.Routing" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.IISIntegration" Version="1.1.4" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="1.1.3" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging.TraceSource" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="1.1.2" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="1.1.3" />
     <PackageReference Include="WindowsAzure.Storage" Version="8.1.4" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Nerdbank.GitVersioning" Version="1.6.35" PrivateAssets="all" />
-    <PackageReference Include="BundlerMinifier.Core" Version="2.2.306" />
-    <DotNetCliToolReference Include="BundlerMinifier.Core" Version="2.2.301" />
+    <PackageReference Include="Nerdbank.GitVersioning" Version="3.4.194" PrivateAssets="all" />
+    <PackageReference Include="tinySite" Version="0.2.8" PrivateAssets="all" />
   </ItemGroup>
 
-  <Target Name="PrecompileScript" BeforeTargets="BeforeBuild">
-    <Exec Command="..\..\tools\tinysite\tinysite.exe render Static" />
-
+  <Target Name="PrecompileScript" BeforeTargets="BeforeBuild"
+          DependsOnTargets="BuildSite">
     <ItemGroup>
-      <Content Remove="wwwroot/**" />
-      <Content Include="wwwroot/**" CopyToPublishDirectory="PreserveNewest" CopyToOutputDirectory="PreserveNewest" />
+      <Content Update="wwwroot/**" CopyToPublishDirectory="PreserveNewest" CopyToOutputDirectory="PreserveNewest" />
     </ItemGroup>
   </Target>
-
 </Project>
